Allowed parsing expressions as statements
This commit modifies the `StatementParser` to allow expressions to be parsed as statements.

Previously, the parser would throw an `UnexpectedTokenException` if it encountered an expression where a statement was expected. Now, if other statement parsing attempts fail, the parser will try to parse the input as an expression and wrap it in an `ExpressionStatement`.

This change introduces a new `ExpressionStatement` AST node to represent such cases.